refactor: 이벤트 핸들러 함수들을 eventHandlers 디렉터리로 이동
